+++ plugins/vscode/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the Viv VS Code extension (`Viv DSL`) will be documented 
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.10.6] – 2026-04-10
+
+### Changed
+
+* Bump compatible compiler version to `0.11`.
+
 ## [0.10.5] – 2026-04-07
 
 ### Changed